feat(ir): Add memory address allocation for non-DDR spaces in AddAllocPass (#103)
## Summary

- **AddAllocPass**: Allocate non-overlapping 32-byte aligned addresses
for `MemRef`s in non-DDR memory spaces (UB, L1, L0A, L0B, L0C). DDR
`MemRef`s keep their original addresses.
- Group `MemRef`s by memory space and assign sequential addresses per
space.
- Update all `MemRef` references in the IR with the new allocated
addresses.
- Emit alloc statements in address order.
- **IR / serialization**: ConstInt structural equality, hash, and
serialization support (scalar_expr, serializer, type_deserializers,
structural_equal, structural_hash).
- **Tests**: Add tests for address alignment and allocation in
`test_add_alloc_pass.py`, ConstInt in `test_constants.py`; update
equality tests for new error message format.

## Changes

| Area | Description |
|------|-------------|
| `add_alloc_pass.cpp` | Memory-space grouping, 32-byte alignment,
address allocation and IR update |
| `scalar_expr.h` / `ir.cpp` | ConstInt-related binding/API |
| Serialization | ConstInt in serializer and type_deserializers |
| structural_equal / structural_hash | ConstInt support and error
message format |
| Tests | AddAllocPass, constants, and equality regex updates |

## Testing

- [x] Pre-commit passes (clang-format, cpplint, ruff, pyright)
- [x] `python3 -m pytest . -v -s` passes
- [x] AddAllocPass tests and ConstInt/equality tests updated

## Related

- Branch: `mem-alloc`

+++ include/pypto/ir/scalar_expr.h
@@ -76,7 +76,7 @@ using ScalarExprPtr = std::shared_ptr<const ScalarExpr>;
  */
 class ConstInt : public Expr {
  public:
-  const int value_;  // Numeric constant value (immutable)
+  const int64_t value_;  // Numeric constant value (immutable)
 
   /**
    * @brief Create a constant expression
@@ -84,7 +84,7 @@ class ConstInt : public Expr {
    * @param value Numeric value
    * @param span Source location
    */
-  ConstInt(int value, DataType dtype, Span span)
+  ConstInt(int64_t value, DataType dtype, Span span)
       : Expr(std::move(span), std::make_shared<ScalarType>(dtype)), value_(value) {}
 
   [[nodiscard]] ObjectKind GetKind() const override { return ObjectKind::ConstInt; }
